Move actions to their own directory.
