+++ mapper/actions/action.js
@@ -1,0 +1,14 @@
+class Action {
+	constructor(context, options) {
+		this.context = context;
+		this.options = options;
+	}
+
+	empty() {
+		return true;
+	}
+
+	async perform() {}
+}
+
+export { Action };
