+++ mapper/actions/index.js
@@ -1,0 +1,7 @@
+export * from "./action.js";
+export * from "./bulk_action.js";
+export * from "./draw_path_action.js";
+export * from "./node_cleanup_action.js";
+export * from "./remove_action.js";
+export * from "./translate_action.js";
+export * from "./unremove_action.js";
