+++ mapper/mapper.js
@@ -1,187 +1,7 @@
 import { HookContainer } from "./hook_container.js";
 import { Vector3, Box3, Line3 } from "./geometry.js";
 import { asyncFrom, mod } from "./utils.js";
-
-class Action {
-	constructor(context, options) {
-		this.context = context;
-		this.options = options;
-	}
-
-	empty() {
-		return true;
-	}
-
-	async perform() {}
-}
-
-class DrawPathAction extends Action {
-	async perform() {
-		const path = this.options.path;
-		const scrollOffset = this.options.scrollOffset;
-
-		const pathOnMap = path.mapOrigin((origin) => origin.add(scrollOffset)).withBisectedLines(this.options.radius);
-
-		const placedNodes = [];
-
-		if(!this.options.parent) {
-			this.options.parent = await this.context.mapper.insertNode(pathOnMap.getCenter(), {
-				type: this.options.nodeType,
-				radius: 0,
-			});
-
-			placedNodes.push(this.options.parent);
-		}
-
-		const pathCenter = pathOnMap.getCenter();
-
-		const vertices = Array.from(pathOnMap.vertices()).sort((a, b) => a.subtract(pathCenter).lengthSquared() - b.subtract(pathCenter).lengthSquared());
-
-		const placedVertices = [];
-
-		const radius = this.options.radius;
-
-		placeEachVertex: for(const vertex of vertices) {
-			for(const placedVertex of placedVertices) {
-				if(placedVertex.point.subtract(vertex).length() < (radius + placedVertex.radius) / 4) {
-					continue placeEachVertex;
-				}
-			}
-
-			if(radius > 0) {
-				placedNodes.push(await this.context.mapper.insertNode(vertex, {
-					type: this.options.nodeType,
-					radius: radius,
-					parent: this.options.parent,
-				}));
-
-				placedVertices.push({
-					point: vertex,
-					radius: radius,
-				});
-			}
-		}
-
-		const undoAction = new RemoveAction(this.context, {
-			nodeRefs: placedNodes,
-		});
-
-		if(this.options.fullCalculation) {
-			const undoCleanupAction = await this.context.performAction(new NodeCleanupAction(this.context, {nodeRef: this.options.parent, type: this.options.nodeType}), false);
-			return new BulkAction(this.context, {
-				actions: [undoCleanupAction, undoAction],
-			});
-		}
-		else {
-			return undoAction;
-		}
-	}
-
-	empty() {
-		return false;
-	}
-}
-
-class NodeCleanupAction extends Action {
-	async perform() {
-		const toRemove = new Set();
-
-		const vertices = (await asyncFrom(this.getAllVertices())).sort((a, b) => a.radius - b.radius);
-
-		for(const vertex of vertices) {
-			if(vertex.removable) {
-				for(const otherVertex of vertices) {
-					if(!toRemove.has(otherVertex.nodeRef.id) && otherVertex !== vertex && otherVertex.point.subtract(vertex.point).length() < (vertex.radius + otherVertex.radius) / 4) {
-						toRemove.add(vertex.nodeRef.id);
-					}
-				}
-			}
-		}
-
-		return await this.context.performAction(new RemoveAction(this.context, {nodeRefs: [...toRemove].map((id) => this.context.mapper.backend.getNodeRef(id))}), false);
-	}
-
-	async * getAllNodes() {
-		for await (const nodeRef of this.options.nodeRef.getSelfAndAllDescendants()) {
-			if(await nodeRef.getPString("type") === this.options.type) {
-				yield nodeRef;
-			}
-		}
-	}
-
-	async * getAllVertices() {
-		for await (const nodeRef of this.getAllNodes()) {
-			const radius = await nodeRef.getPNumber("radius");
-			if(radius > 0) {
-				yield {
-					nodeRef: nodeRef,
-					removable: !(await nodeRef.hasChildren()),
-					point: await nodeRef.center(),
-					radius: radius,
-				};
-			}
-		}
-	}
-}
-
-class RemoveAction extends Action {
-	async perform() {
-		const affectedNodeRefs = await this.context.mapper.removeNodes(this.options.nodeRefs);
-		return new UnremoveAction(this.context, {nodeRefs: affectedNodeRefs});
-	}
-
-	empty() {
-		return this.options.nodeRefs.length === 0;
-	}
-}
-
-class UnremoveAction extends Action {
-	async perform() {
-		await this.context.mapper.unremoveNodes(this.options.nodeRefs);
-		return new RemoveAction(this.context, {nodeRefs: this.options.nodeRefs});
-	}
-
-	empty() {
-		return this.options.nodeRefs.length === 0;
-	}
-}
-
-class TranslateAction extends Action {
-	async perform() {
-		await this.context.mapper.translateNode(this.options.nodeRef, this.options.offset);
-		return new TranslateAction(this.context, {
-			nodeRef: this.options.nodeRef,
-			offset: this.options.offset.multiplyScalar(-1),
-		});
-	}
-
-	empty() {
-		return false;
-	}
-}
-
-class BulkAction extends Action {
-	async perform() {
-		const actions = [];
-
-		for(const action of this.options.actions.reverse()) {
-			actions.push(await this.context.performAction(action, false));
-		}
-
-		return new BulkAction(this.context, {
-			actions: actions,
-		});
-	}
-
-	empty() {
-		for(const action of this.options.actions) {
-			if(!action.empty()) {
-				return false;
-			}
-		}
-		return true;
-	}
-}
+import { BulkAction, DrawPathAction, TranslateAction, RemoveAction } from "./actions/index.js";
 
 class Brush {
 	constructor(context) {
